norm (Excluding parsecmd_tokenize.c files), fix : exit_status
